+++ bot/discord/src/jvmMain/kotlin/io/github/sophon/discord/domain/Tracker.kt
@@ -37,6 +37,7 @@ internal interface Tracker {
         featureName: String,
         command: BotCommand,
     ): EmptyResult<BotError>
+    suspend fun recordFailure(): EmptyResult<BotError>
 }
 
 @OptIn(ExperimentalTime::class)
@@ -74,6 +75,16 @@ internal class TrackerImpl(
         return statsTracker.record(trackedCommand)
             .mapError { it.toDomainError() }
             .onError { Napier.e(tag = TAG) { it.toString() } }
+    }
+
+    override suspend fun recordFailure(): EmptyResult<BotError> {
+        val failure = TrackedCommand(
+            feature = "failed",
+            name = "failed",
+        )
+
+        return statsTracker.record(failure)
+            .mapError { it.toDomainError() }
     }